feat: move ton rate to header
